// packages/twenty-server/src/metadata/field-metadata/interfaces/relation-metadata.interface.ts
import { RelationMetadataType } from 'src/metadata/relation-metadata/relation-metadata.entity';

import { ObjectMetadataInterface } from './object-metadata.interface';
import { FieldMetadataInterface } from './field-metadata.interface';

export interface RelationMetadataInterface {
  id: string;

  relationType: RelationMetadataType;

  fromObjectMetadataId: string;
  fromObjectMetadata: ObjectMetadataInterface;

  toObjectMetadataId: string;
  toObjectMetadata: ObjectMetadataInterface;

  fromFieldMetadataId: string;
  fromFieldMetadata: FieldMetadataInterface;

  toFieldMetadataId: string;
  toFieldMetadata: FieldMetadataInterface;
}